Fix gender bug and improve UI
The gender bug was because the slider variables were stored as a
string instead of a number. This meant, that the gender variable is
always truthy.
I fixed this by using parseFloat to force the value saved to be numeric.

I changed the UI a bit to match the design more.