Fix object order problem
There has been several changes in this commit.
The main change is now the slider controls the intended model.
Before only the right lung was controlled.
The reason was because of async operations putting the objects
stored in a temp array out of order.
To fix this I initialised the array and made sure the objects saved
were in the correct order.

Another minor change is the removal of transparency controls.
I'll do something about switching transparency later on.